Fix bug causing false positive hotspot activations on button release
When a mouse button was pressed over a hotspot, but released over none, the
press_hotspot was not cleared. I now the cursor was dragged over the hotspot
later while pressed, withouth wayprompt receiving the press event (like when
the button is pressed outside our surface), and then released, it caused a
false positive activation of the hotspot action.